cambios

+++ auth-service/src/main/java/com/sakila/auth/repository/StaffRepository.java
@@ -2,8 +2,9 @@ package com.sakila.auth.repository;
 
 import com.sakila.auth.entity.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface StaffRepository extends JpaRepository<Staff, Integer> {
+public interface StaffRepository extends JpaRepository<Staff, Byte> {
     Optional<Staff> findByUsername(String username);
 }
