+++ business-service/src/main/java/com/sakila/business/repository/StoreRepository.java
@@ -3,5 +3,5 @@ package com.sakila.business.repository;
 import com.sakila.business.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Byte> {
 }
