+++ inventory-service/src/main/java/com/sakila/inventory/repository/FilmRepository.java
@@ -3,5 +3,5 @@ package com.sakila.inventory.repository;
 import com.sakila.inventory.entity.Film;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmRepository extends JpaRepository<Film, Integer> {
+public interface FilmRepository extends JpaRepository<Film, Short> {
 }
